Refactor to reduce number of functions
